Make url parameter detection more robust
The previous solution assumed that there is at most one URL parameter when
importing an SVG. This assumption could break in certain scenarios e.g. when
using vite-plugin-ssr and import.meta.glob with a "?url" parameter.

The new solution will use the builtin URLSearchParams and cycle through all
parameters until it finds one that it knows, otherwise the default is used.